Restructuring code #162
